Added the initial version of the To-do view

+++ lib/utils/border_style.dart
@@ -3,7 +3,7 @@ import 'package:gc_clone/constants.dart';
 
 BoxDecoration kBorderStyle = BoxDecoration(
   borderRadius: BorderRadius.circular(
-    kBorderCurver as double,
+    kBorderCurver,
   ),
   border: Border.all(
     color: kTextColor ?? Colors.grey,
